pyproject.toml
[project]
name = "bqcli"
version = "0.1.0"
description = "REPL for BigQuery"
authors = [
    {name = "Jan Červenka", email = "jan.cervenka@yahoo.com"},
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-bigquery>=3.27.0",
    "pandas>=2.2.3",
    "prompt-toolkit>=3.0.48",
    "pygments>=2.18.0",
]

[dependency-groups]
dev = [
    "pexpect>=4.9.0",
    "pytest>=8.3.4",
]

[project.scripts]
example-pkg = "bqcli:repl"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"